Исправлена ячейка на один класс
