+++ TinkoffChat/TinkoffChat/Conversation/ConversationMessageCell.xib
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina4_0" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" restorationIdentifier="ConversationMessageCell" selectionStyle="default" indentationWidth="10" reuseIdentifier="ConversationMessageCell" id="KGk-i7-Jjw" customClass="ConversationMessageCell" customModule="TinkoffChat" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+                <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="AkB-H5-kpB">
+                        <rect key="frame" x="8" y="8" width="58" height="28"/>
+                        <subviews>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7xR-b1-JaZ">
+                                <rect key="frame" x="8" y="8" width="42" height="12"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstAttribute="trailing" secondItem="7xR-b1-JaZ" secondAttribute="trailing" constant="8" id="2q4-Z6-WaA"/>
+                            <constraint firstItem="7xR-b1-JaZ" firstAttribute="leading" secondItem="AkB-H5-kpB" secondAttribute="leading" constant="8" id="F6U-jA-522"/>
+                            <constraint firstItem="7xR-b1-JaZ" firstAttribute="top" secondItem="AkB-H5-kpB" secondAttribute="top" constant="8" id="Fhe-uu-8ec"/>
+                            <constraint firstAttribute="bottom" secondItem="7xR-b1-JaZ" secondAttribute="bottom" constant="8" id="M0S-12-FeN"/>
+                        </constraints>
+                    </view>
+                </subviews>
+                <color key="backgroundColor" red="0.83741801979999997" green="0.83743780850000005" blue="0.83742713930000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                <constraints>
+                    <constraint firstItem="7xR-b1-JaZ" firstAttribute="width" relation="lessThanOrEqual" secondItem="H2p-sc-9uM" secondAttribute="width" multiplier="7:10" id="9iR-xm-QON"/>
+                    <constraint firstAttribute="bottom" secondItem="AkB-H5-kpB" secondAttribute="bottom" constant="8" id="Lu0-hS-kWF"/>
+                    <constraint firstItem="AkB-H5-kpB" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="8" id="nvc-Pp-s2Q"/>
+                    <constraint firstItem="AkB-H5-kpB" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="8" id="vqA-sP-sW0"/>
+                    <constraint firstAttribute="trailing" relation="lessThanOrEqual" secondItem="AkB-H5-kpB" secondAttribute="trailing" constant="254" id="yrE-Pq-jSg"/>
+                </constraints>
+            </tableViewCellContentView>
+            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
+            <connections>
+                <outlet property="backgroundMessageView" destination="AkB-H5-kpB" id="HV7-ga-IPN"/>
+                <outlet property="leftMessageConstraint" destination="nvc-Pp-s2Q" id="Not-WX-WBb"/>
+                <outlet property="messageLabel" destination="7xR-b1-JaZ" id="kEZ-Iv-v4r"/>
+                <outlet property="rightMessageConstraint" destination="yrE-Pq-jSg" id="TD8-hh-HuT"/>
+            </connections>
+            <point key="canvasLocation" x="131.25" y="109.85915492957747"/>
+        </tableViewCell>
+    </objects>
+</document>
