--- TinkoffChat/TinkoffChat/Conversation/ConversationOutgoingMessageCell.xib
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_0" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" id="KGk-i7-Jjw" customClass="ConversationOutgoingMessageCell" customModule="TinkoffChat" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="320" height="44"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="lMR-WT-FBP">
-                        <rect key="frame" x="254" y="8" width="58" height="28"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                    </view>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="right" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="dN8-d4-PZi">
-                        <rect key="frame" x="262" y="16" width="42" height="12"/>
-                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                </subviews>
-                <color key="backgroundColor" red="0.83741801979999997" green="0.83743780850000005" blue="0.83742713930000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                <constraints>
-                    <constraint firstItem="dN8-d4-PZi" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="16" id="2Es-mk-QUZ"/>
-                    <constraint firstAttribute="bottom" secondItem="dN8-d4-PZi" secondAttribute="bottom" constant="16" id="6yi-vC-5Pu"/>
-                    <constraint firstItem="lMR-WT-FBP" firstAttribute="leading" secondItem="dN8-d4-PZi" secondAttribute="leading" constant="-8" id="FCG-hP-K9v"/>
-                    <constraint firstItem="lMR-WT-FBP" firstAttribute="trailing" secondItem="dN8-d4-PZi" secondAttribute="trailing" constant="8" id="I4z-jE-Hlg"/>
-                    <constraint firstItem="dN8-d4-PZi" firstAttribute="width" relation="lessThanOrEqual" secondItem="H2p-sc-9uM" secondAttribute="width" multiplier="7:10" id="TRi-ai-qzy"/>
-                    <constraint firstAttribute="trailing" secondItem="dN8-d4-PZi" secondAttribute="trailing" constant="16" id="jBj-cF-QbN"/>
-                    <constraint firstItem="lMR-WT-FBP" firstAttribute="top" secondItem="dN8-d4-PZi" secondAttribute="top" constant="-8" id="jhn-uD-kOr"/>
-                    <constraint firstItem="lMR-WT-FBP" firstAttribute="bottom" secondItem="dN8-d4-PZi" secondAttribute="bottom" constant="8" id="kgt-hj-egA"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="backgroundMessageView" destination="lMR-WT-FBP" id="tiX-v9-gv4"/>
-                <outlet property="outgoingMessageLabel" destination="dN8-d4-PZi" id="DzN-XT-Is5"/>
-            </connections>
-            <point key="canvasLocation" x="131.8840579710145" y="150"/>
-        </tableViewCell>
-    </objects>
-</document>
